requirements.txt
Pillow==9.2.0
pip==22.2.2
setuptools==60.2.0
wheel==0.37.1
